% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ucrt.dp01.cal.rh.dew.frst.pt.R
\name{def.ucrt.dp01.cal.rh.dew.frst.pt}
\alias{def.ucrt.dp01.cal.rh.dew.frst.pt}
\title{Compute L1 calibration (temporally aggregated) uncertainty of a dew point measurement}
\usage{
def.ucrt.dp01.cal.rh.dew.frst.pt(data, VarUcrt, ucrtData, log = NULL)
}
\arguments{
\item{data}{Vector}

\item{VarUcrt}{A character string of the name of the dew point variable.}

\item{ucrtData}{A data frame of relevant L0' individual measurement uncertainty data for 
dew point measurement from NEONprocIS.cal::def.ucrt.meas.rh.dew.frst.pt. 
Columns must include:\cr
readout_time (POSIX),\cr
VarUcrt_ucrt_dfpt_t_L1 (numeric) - 1st derivative of calibration function with respect to temperature multiplied by the L1 uncertainty coefficient, and\cr 
VarUcrt_ucrt_dfpt_rh_L1 (numeric) - 1st derivative of calibration function with respect to relative humidity multiplied by the L1 uncertainty coefficient, and\cr 
where VarUcrt is the variable name specified in input \code{VarUcrt}.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A data frame of the L1 dew point calibration uncertainties with respect to temperature and relative humidity.
}
\description{
Definition function. Compute the L1 calibration (temporally aggregated) uncertainty of a
dew point measurement
}
\examples{
data <- c(10,11,9)
ucrtData <- data.frame(readout_time=as.POSIXct(c('2019-01-01 00:00','2019-01-01 00:01','2019-01-01 00:02'),tz='GMT'),
                       dewPoint_ucrt_dfpt_t_L1=c(0.05,0.06,0.055),
                       dewPoint_ucrt_dfpt_rh_L1=c(0.065,0.055,0.054),
                       stringsAsFactors=FALSE)
ucrtCal <- NEONprocIS.stat::def.ucrt.dp01.cal.rh.dew.frst.pt(data=data,VarUcrt='dewPoint',ucrtData=ucrtData)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
NEON.DOC.000851 NEON ATBD Humidity and temperature
}
\seealso{
\link[NEONprocIS.stat]{wrap.ucrt.dp01}

\link[NEONprocIS.cal]{def.ucrt.meas.rh.dew.frst.pt}
}
\author{
Edward Ayres \email{eayres@battelleecology.org}
}
\keyword{L1,}
\keyword{average}
\keyword{calibration,}
\keyword{dew}
\keyword{point,}
\keyword{uncertainty,}
