% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.ucrt.dp01.cal.cnst.R
\name{wrap.ucrt.dp01.cal.cnst}
\alias{wrap.ucrt.dp01.cal.cnst}
\title{Compute combined and expanded (95% confidence) temporally aggregated L1 uncertainty 
due to natural variation and calibration (constant value)}
\usage{
wrap.ucrt.dp01.cal.cnst(data, VarUcrt, ucrtCoef, ucrtData = NULL, log = NULL)
}
\arguments{
\item{data}{Data frame of L0' (calibrated) data. Must contain columns \code{readout_time} (POSIX) and 
whatever variable is specified in input parameter \code{VarUcrt} (numeric).
A single aggregated uncertainty for the selected variable \code{VarUcrt} will be computed over the full timeseries.}

\item{VarUcrt}{A character string of the target variable (column) in the data frame \code{data} for 
which uncertainty data will be computed (all other columns will be ignored in this function).}

\item{ucrtCoef}{A list of uncertainty coefficients, each a list containing at a minimum the list 
elements: term (name of L0' term for which the coefficient applies - string), start_date (POSIX), 
end_date(POSIX), Name (of the coefficient - string), and 
Value (of the coefficient - string or numeric, to be interpreted as numeric). 
This will be passed into the calibration uncertainty function. Calibration uncertainty 
requires the U_CVALA3 coefficient.}

\item{ucrtData}{Unused in this function.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A single numeric value representing the aggregated L1 calibration uncertainty over the full record. 
Numeric NA is returned if no applicable coefficient is found, with warning.
}
\description{
Wrapper function. Compute the combined and expanded (95% confidence) temporally aggregated L1 
uncertainty for a set of values subject to natural variation and calibration uncertainty. 
Uncertainty due to  natural variation is estimated from the 
standard error of the mean. Uncertainty due to calibration is represented by a constant value 
coefficient provided by CVAL (U_CVALA3).
}
\examples{
data <- data.frame(readout_time=as.POSIXct(c('2019-01-01 00:00','2019-01-01 00:01','2019-01-01 00:02'),tz='GMT'),
                   temp=c(.599,.598,.597))
ucrtCoef <- list(list(term='temp',start_date=as.POSIXct('2019-01-01',tz='GMT'),end_date=as.POSIXct('2019-01-02',tz='GMT'),Name='U_CVALA3',Value='0.0141'))
ucrtData <- data.frame(readout_time=as.POSIXct(c('2019-01-01 00:00','2019-01-01 00:01','2019-01-01 00:02'),tz='GMT'),
                       temp=c(100.187,100.195,100.203))
ucrt <- NEONprocIS.stat::wrap.ucrt.dp01.cal.cnst(data=data,VarUcrt='temp',ucrtCoef=ucrtCoef)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
}
\seealso{
\link[NEONprocIS.stat]{wrap.ucrt.dp01}

\link[NEONprocIS.stat]{def.ucrt.dp01.cal.cnst}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{L1,}
\keyword{average}
\keyword{calibration,}
\keyword{fdas}
\keyword{uncertainty,}
