% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.ucrt.dp01.cal.mult.fdas.volt.R
\name{wrap.ucrt.dp01.cal.mult.fdas.volt}
\alias{wrap.ucrt.dp01.cal.mult.fdas.volt}
\title{Compute combined and expanded (95% confidence) temporally aggregated L1 uncertainty 
due to natural variation, calibration (multiplier), and voltage-based FDAS measurement}
\usage{
wrap.ucrt.dp01.cal.mult.fdas.volt(
  data,
  VarUcrt,
  ucrtCoef,
  ucrtData,
  log = NULL
)
}
\arguments{
\item{data}{Data frame of L0' (calibrated) data. Must contain columns \code{readout_time} (POSIX) and 
whatever variable is specified in input parameter \code{VarUcrt} (numeric).
A single aggregated uncertainty for the selected variable \code{VarUcrt} will be computed over the full timeseries.}

\item{VarUcrt}{A character string of the target variable (column) in the data frame \code{data} for 
which uncertainty data will be computed (all other columns will be ignored in this function).}

\item{ucrtCoef}{A list of uncertainty coefficients, each a list containing at a minimum the list 
elements: term (name of L0' term for which the coefficient applies - string), start_date (POSIX), 
end_date(POSIX), Name (of the coefficient - string), and 
Value (of the coefficient - string or numeric, to be interpreted as numeric). 
This will be passed into the calibration and FDAS uncertainty functions. Calibration uncertainty 
requires the U_CVALA3 coefficient. Voltage-based FDAS uncertainty requires U_CVALV3 and U_CVALV4 
coefficients.}

\item{ucrtData}{A data frame of L0' individual measurement uncertainty data generated by the 
calibration step. This will be passed into the FDAS uncertainty function NEONprocIS.stat::def.ucrt.dp01.fdas. 
See that function's inputs for required columns.}

\item{log}{A logger object as produced by NEONprocIS.base::def.log.init to produce structured log
output. Defaults to NULL, in which the logger will be created and used within the function.}
}
\value{
A single numeric value representing the aggregated L1 calibration uncertainty over the full record. 
Numeric NA is returned if no applicable coefficient is found, with warning.
}
\description{
Wrapper function. Compute the combined and expanded (95% confidence) temporally aggregated L1 
uncertainty for a set of values subject to natural variation, calibration uncertainty, and 
voltage-based FDAS uncertainty. Uncertainty due to  natural variation is estimated from the 
standard error of the mean. Uncertainty due to calibration is represented by a multiplier 
(U_CVALA3 (%) coefficient provided by CVAL) to the maximum L0' value. Uncertainty due to 
voltage-based FDAS measurement is computed from NEONprocIS.stat::def.ucrt.dp01.fdas.
}
\examples{
data <- data.frame(readout_time=as.POSIXct(c('2019-01-01 00:00','2019-01-01 00:01','2019-01-01 00:02'),tz='GMT'),
                   linePAR=c(827.7,827.8,831.3))
ucrtCoef <- list(list(term='linePAR',start_date=as.POSIXct('2019-01-01',tz='GMT'),end_date=as.POSIXct('2019-01-02',tz='GMT'),Name='U_CVALA3',Value='0.0388'),
                 list(term='linePAR',start_date=as.POSIXct('2019-01-01',tz='GMT'),end_date=as.POSIXct('2019-01-02',tz='GMT'),Name='U_CVALV3',Value='0.000374'),
                 list(term='linePAR',start_date=as.POSIXct('2019-01-01',tz='GMT'),end_date=as.POSIXct('2019-01-02',tz='GMT'),Name='U_CVALV4',Value='0.0000020'))
ucrtData <- data.frame(readout_time=as.POSIXct(c('2019-01-01 00:00','2019-01-01 00:01','2019-01-01 00:02'),tz='GMT'),
                       linePAR_raw=c(0.0064085,0.006409,0.0064097),
                       linePAR_dervCal=c(129153.7,129153.7,129153.7),
                       linePAR_ucrtComb=c(0.57992,0.57998,0.57998),
                       stringsAsFactors=FALSE)
ucrt <- NEONprocIS.stat::wrap.ucrt.dp01.cal.mult.fdas.volt(data=data,VarUcrt='linePAR',ucrtCoef=ucrtCoef,ucrtData=ucrtData)
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
NEON.DOC.000785 TIS Level 1 Data products Uncertainty Budget Estimation Plan
NEON.DOC.000746 Calibration Fixture and Sensor Uncertainty Analysis: CVAL 2014 Uncertainty Manual
}
\seealso{
\link[NEONprocIS.stat]{wrap.ucrt.dp01}

\link[NEONprocIS.stat]{def.ucrt.dp01.cal.mult}

\link[NEONprocIS.stat]{def.ucrt.dp01.fdas}
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{L1,}
\keyword{average}
\keyword{calibration,}
\keyword{fdas}
\keyword{uncertainty,}
